Move deleting picture files into a module, delete pictures when picture is deleted
